Squashed commit of the following:

commit d402a299c4168f67eea200fc0e5973071a6bc5c1
Author: Oleksii Khomiakovskyi <okhomiak@cisco.com>
Date:   Mon Jan 20 16:03:04 2020 +0200

    build: fix build on big-endian systems

    moved little-endian arrays conversion macros in a common header
    added macros to convert little endian to host order for unaligned access
    made unit tests independent of types size and system arch

+++ src/flow/test/flow_cache_test.cc
@@ -88,8 +88,8 @@ bool HighAvailabilityManager::in_standby(Flow*) { return true; }
 SfIpRet SfIp::set(void const*, int) { return SFIP_SUCCESS; }
 namespace memory
 {
-void MemoryCap::update_allocations(unsigned long) { }
-void MemoryCap::update_deallocations(unsigned long) { }
+void MemoryCap::update_allocations(size_t) { }
+void MemoryCap::update_deallocations(size_t) { }
 bool MemoryCap::over_threshold() { return true; }
 }
